+++ src/main/kotlin/step3/racing/printer/CarPrinterImpl.kt
@@ -1,7 +1,7 @@
 package step3.racing.printer
 
 import step3.racing.car.Car
-import step3.racing.mock.View
+import step3.racing.view.View
 
 class CarPrinterImpl(override val view: View) : CarPrinter {
     override fun printString(car: Car) = (0 until car.distance)
