+++ src/test/kotlin/step3/racing/view/MockTurn.kt
@@ -1,4 +1,4 @@
-package step3.racing.mock
+package step3.racing.view
 
 import step3.turn.Turn
